[Refactor] #14 ErrorCode 인터페이스로 선언
