+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/global/config/swagger/SwaggerResponseDescription.java
@@ -1,12 +1,13 @@
 package org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.config.swagger;
 
 import lombok.Getter;
-import org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.ErrorCode;
+import org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.constant.ErrorCode;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-import static org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.ErrorCode.*;
+import static org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.constant.GlobalErrorCode.*;
+import static org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.constant.ProductErrorCode.PRODUCT_NOT_FOUND;
 
 @Getter
 public enum SwaggerResponseDescription {
