+++ src/main/java/org/sopt/SOPT_36_COLLABORATION_SERVER_TEMU/global/exception/GlobalExceptionHandler.java
@@ -4,7 +4,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.response.BaseErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
-import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -15,9 +14,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.NoHandlerFoundException;
 
-import java.util.Objects;
-
-import static org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.ErrorCode.*;
+import static org.sopt.SOPT_36_COLLABORATION_SERVER_TEMU.global.exception.constant.GlobalErrorCode.*;
 
 @Slf4j
 @RestControllerAdvice
